Add editor option to write logs with text editor
When given the option -e or --editor <editor>, ql will now open a
temporary file into a provided text editor and use the written contents
of the temporary file to formulate the log. The log will be formatted
according to the same rules as when given directly to ql as separate
arguments receiving a HH:mm timestamp and tab indentation.

Separate write logic to own function `write_log`, which receives either
the temporary log file content or the provided arguments.